requirements.txt
python-dotenv~=1.1.0
langchain~=0.3.25
langchain-openai~=0.3.21
langchain-groq
langchain-google-genai~=2.1.6
langchain-anthropic~=0.3.15
langchain-xai
langchain-core~=0.3.64
langdetect~=1.0.9
